Adição do método editar e dos erros do ApiResponse

+++ src/main/java/br/ufsm/poow2/biblioteca_rest/model/Genero.java
@@ -15,7 +15,7 @@ public class Genero {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idGenero;
+    private Integer idGenero;
 
     @Column(name = "nomeGenero")
     private @NotBlank String nomeGenero;
